initial refactorings related to CSIG
- separated mvf-core.* into globals and functions which are only used by the main programs. This is so that the extended signature hashing functionality (extended with chainId) can be used in non-main executables (bitcoin-tx, possibly others later).
- ForkSetup has been made into a boolean and the StartShutdown() calls factored out to remove this additional dependency on main program code.
- mvf-core-globals are now linked into libbitcoin_common , and so accessible to all programs, even those which do not require extended fork processing routines.
