+++ src/Makefile.am
@@ -118,6 +118,7 @@ BITCOIN_CORE_H = \
   merkleblock.h \
   miner.h \
   mvf-core.h \
+  mvf-core-globals.h \
   net.h \
   netbase.h \
   noui.h \
@@ -199,6 +200,7 @@ libbitcoin_server_a_SOURCES = \
   merkleblock.cpp \
   miner.cpp \
   mvf-core.cpp \
+  mvf-core-globals.cpp \
   net.cpp \
   noui.cpp \
   policy/fees.cpp \
@@ -282,6 +284,7 @@ libbitcoin_common_a_SOURCES = \
   key.cpp \
   keystore.cpp \
   netbase.cpp \
+  mvf-core-globals.cpp \
   primitives/block.cpp \
   primitives/transaction.cpp \
   protocol.cpp \
